Moving fixture into separate module
